refactor(eswrapper): rename dir to match package name per review
Per @yurishkuro on #8386: rather than renaming the package from eswrapper
to wrapper to match the directory, rename the directory from wrapper/ to
eswrapper/. Callers already imported it as eswrapper, so this removes the
import alias and keeps the more informative package name.

Also applies Copilot feedback: update the comment in .golangci.yml to
reflect the current path-based exclusion strategy for storage v1/v2 APIs.
