+++ internal/storage/elasticsearch/config/config.go
@@ -35,7 +35,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/auth"
 	"github.com/jaegertracing/jaeger/internal/metrics"
 	es "github.com/jaegertracing/jaeger/internal/storage/elasticsearch"
-	eswrapper "github.com/jaegertracing/jaeger/internal/storage/elasticsearch/wrapper"
+	"github.com/jaegertracing/jaeger/internal/storage/elasticsearch/eswrapper"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore/spanstoremetrics"
 )
