+++ internal/storage/elasticsearch/eswrapper/wrapper.go
@@ -2,7 +2,7 @@
 // Copyright (c) 2017 Uber Technologies, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package wrapper
+package eswrapper
 
 import (
 	"context"
