+++ internal/storage/elasticsearch/eswrapper/wrapper_nolint.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package wrapper
+package eswrapper
 
 import es "github.com/jaegertracing/jaeger/internal/storage/elasticsearch"
